+++ app/models/concerns/deletable.rb
@@ -1,3 +1,4 @@
+# Allows database rows to be scoped and marked as deleted
 module Deletable
   extend ActiveSupport::Concern