+++ app/models/user.rb
@@ -10,6 +10,7 @@ class User < ActiveRecord::Base
   has_many :accounts, -> { where deleted: false }
   has_many :charge_types, -> { where deleted: false }, through: :accounts
   has_many :entries, -> { where deleted: false }
+  has_many :templates, -> { current }
 
   def total_expenditures
     @total_expenditures ||= begin