+++ config/routes.rb
@@ -23,12 +23,13 @@ Rails.application.routes.draw do
     end
   end
 
-  devise_for :users,  controllers: { registrations: 'contour/registrations',
-                                          sessions: 'contour/sessions',
-                                         passwords: 'contour/passwords' },
-                      path_names:  {       sign_up: 'register',
-                                           sign_in: 'login' },
-                      path: ''
+  resources :templates
+
+  devise_for :users, controllers: { registrations: 'contour/registrations',
+                                    sessions: 'contour/sessions',
+                                    passwords: 'contour/passwords' },
+                     path_names: { sign_up: 'register', sign_in: 'login' },
+                     path: ''
 
   get '/about' => 'application#about', as: :about
   get '/month' => 'entries#calendar', as: :calendar