+++ db/migrate/20151222050727_create_templates.rb
@@ -1,0 +1,14 @@
+class CreateTemplates < ActiveRecord::Migration
+  def change
+    create_table :templates do |t|
+      t.string :name
+      t.integer :user_id
+      t.boolean :deleted, null: false, default: false
+
+      t.timestamps null: false
+    end
+
+    add_index :templates, :user_id
+    add_index :templates, :deleted
+  end
+end